+++ R/AddCardModule.R
@@ -8,9 +8,29 @@
 #' @export
 add_card_button_ui <- function(id) {
   ns <- shiny::NS(id)
+
+  # Buttons with custom css and
+  # js code to disable the add card button when clicked to prevent multi-clicks
   shiny::tagList(
     shiny::singleton(
       shiny::tags$head(shiny::includeCSS(system.file("css/custom.css", package = "teal.reporter")))
+    ),
+    shiny::singleton(
+      shiny::tags$head(
+        shiny::tags$script(
+          shiny::HTML(
+            sprintf(
+              '
+              $(document).ready(function(event) {
+                $("body").on("click", "#%s", function() {
+                  $(this).addClass("disabled");
+                })
+              })',
+              ns("add_card_ok")
+            )
+          )
+        )
+      )
     ),
     shiny::tags$button(
       id = ns("add_report_card_button"),
@@ -77,7 +97,7 @@ add_card_button_srv <- function(id, reporter, card_fun) {
             placeholder = "Add a comment here...",
             width = "100%"
           ),
-          footer = shiny::tagList(
+          footer = shiny::div(
             shiny::tags$button(
               type = "button",
               class = "btn btn-danger",
@@ -102,6 +122,8 @@ add_card_button_srv <- function(id, reporter, card_fun) {
         shiny::showModal(add_modal())
       })
 
+      # the add card button is disabled when clicked to prevent multi-clicks
+      # please check the ui part for more information
       shiny::observeEvent(input$add_card_ok, {
         card_fun_args_nams <- names(formals(card_fun))
         has_card_arg <- "card" %in% card_fun_args_nams